# Touching inner rings