--- a/src/earcut.js
+++ b/src/earcut.js
@@ -150,6 +150,7 @@ function isEar(ear) {
     while (p !== a) {
         if (p.x >= x0 && p.x <= x1 && p.y >= y0 && p.y <= y1 &&
             pointInTriangle(ax, ay, bx, by, cx, cy, p.x, p.y) &&
+            !equals(a, p) &&
             area(p.prev, p, p.next) >= 0) return false;
         p = p.next;
     }
@@ -182,25 +183,25 @@ function isEarHashed(ear, minX, minY, invSize) {
     // look for points inside the triangle in both directions
     while (p && p.z >= minZ && n && n.z <= maxZ) {
         if (p.x >= x0 && p.x <= x1 && p.y >= y0 && p.y <= y1 && p !== a && p !== c &&
-            pointInTriangle(ax, ay, bx, by, cx, cy, p.x, p.y) && area(p.prev, p, p.next) >= 0) return false;
+        pointInTriangle(ax, ay, bx, by, cx, cy, p.x, p.y, true) && !equals(a, p) && area(p.prev, p, p.next) >= 0) return false;
         p = p.prevZ;
 
         if (n.x >= x0 && n.x <= x1 && n.y >= y0 && n.y <= y1 && n !== a && n !== c &&
-            pointInTriangle(ax, ay, bx, by, cx, cy, n.x, n.y) && area(n.prev, n, n.next) >= 0) return false;
+        pointInTriangle(ax, ay, bx, by, cx, cy, n.x, n.y, true) && !equals(a, n) && area(n.prev, n, n.next) >= 0) return false;
         n = n.nextZ;
     }
 
     // look for remaining points in decreasing z-order
     while (p && p.z >= minZ) {
         if (p.x >= x0 && p.x <= x1 && p.y >= y0 && p.y <= y1 && p !== a && p !== c &&
-            pointInTriangle(ax, ay, bx, by, cx, cy, p.x, p.y) && area(p.prev, p, p.next) >= 0) return false;
+        pointInTriangle(ax, ay, bx, by, cx, cy, p.x, p.y, true) && !equals(a, p) && area(p.prev, p, p.next) >= 0) return false;
         p = p.prevZ;
     }
 
     // look for remaining points in increasing z-order
     while (n && n.z <= maxZ) {
         if (n.x >= x0 && n.x <= x1 && n.y >= y0 && n.y <= y1 && n !== a && n !== c &&
-            pointInTriangle(ax, ay, bx, by, cx, cy, n.x, n.y) && area(n.prev, n, n.next) >= 0) return false;
+        pointInTriangle(ax, ay, bx, by, cx, cy, n.x, n.y, true) && !equals(a, n) && area(n.prev, n, n.next) >= 0) return false;
         n = n.nextZ;
     }
 
@@ -268,7 +269,7 @@ function eliminateHoles(data, holeIndices, outerNode, dim) {
         queue.push(getLeftmost(list));
     }
 
-    queue.sort(compareXThenSlopeWhenEqual);
+    queue.sort(compareXYSlope);
 
     // process holes from left to right
     for (let i = 0; i < queue.length; i++) {
@@ -278,14 +279,17 @@ function eliminateHoles(data, holeIndices, outerNode, dim) {
     return outerNode;
 }
 
-function compareXThenSlopeWhenEqual(a, b) {
-    const result = a.x - b.x;
+function compareXYSlope(a, b) {
+    let result = a.x - b.x;
     // when the left-most point of 2 holes meet at a vertex, sort the holes counterclockwise so that when we find
     // the bridge to the outer shell is always the point that they meet at.
-    if (result === 0 && a.y === b.y) {
-        const aSlope = (a.next.y - a.y) / (a.next.x - a.x);
-        const bSlope = (b.next.y - b.y) / (b.next.x - b.x);
-        return aSlope - bSlope;
+    if (result === 0) {
+        result = a.y - b.y;
+        if (result === 0) {
+            const aSlope = (a.next.y - a.y) / (a.next.x - a.x);
+            const bSlope = (b.next.y - b.y) / (b.next.x - b.x);
+            result = aSlope - bSlope;
+        }
     }
     return result;
 }
@@ -314,8 +318,11 @@ function findHoleBridge(hole, outerNode) {
 
     // find a segment intersected by a ray from the hole's leftmost point to the left;
     // segment's endpoint with lesser x will be potential connection point
+    // unless they intersect at a vertex, then choose the vertex
+    if (equals(hole, p)) return p;
     do {
-        if (hy <= p.y && (hy >= p.next.y || equals(hole, p)) && p.next.y !== p.y) {
+        if (equals(hole, p.next)) return p.next;
+        else if (hy <= p.y && hy >= p.next.y && p.next.y !== p.y) {
             const x = p.x + (hy - p.y) * (p.next.x - p.x) / (p.next.y - p.y);
             if (x <= hx && x >= qx) {
                 qx = x;