--- a/test/test.js
+++ b/test/test.js
@@ -22,20 +22,39 @@ test('empty', () => {
 
 for (const id of Object.keys(expected.triangles)) {
 
-    test(id, () => {
-        const data = flatten(JSON.parse(fs.readFileSync(new URL(`fixtures/${id}.json`, import.meta.url)))),
-            indices = earcut(data.vertices, data.holes, data.dimensions),
-            err = deviation(data.vertices, data.holes, data.dimensions, indices),
-            expectedTriangles = expected.triangles[id],
-            expectedDeviation = expected.errors[id] || 0;
+    for (const rotation of [0, 90, 180, 270]) {
+        test(`${id} rotation ${rotation}`, () => {
+            const coords = JSON.parse(fs.readFileSync(new URL(`fixtures/${id}.json`, import.meta.url)));
+            const theta = rotation * Math.PI / 180;
+            const xx = Math.round(Math.cos(theta));
+            const xy = Math.round(-Math.sin(theta));
+            const yx = Math.round(Math.sin(theta));
+            const yy = Math.round(Math.cos(theta));
+            if (rotation) {
+                for (const ring of coords) {
+                    for (const coord of ring) {
+                        const [x, y] = coord;
+                        coord[0] = xx * x + xy * y;
+                        coord[1] = yx * x + yy * y;
+                    }
+                }
+            }
+            const data = flatten(coords),
+                indices = earcut(data.vertices, data.holes, data.dimensions),
+                err = deviation(data.vertices, data.holes, data.dimensions, indices),
+                expectedTriangles = expected.triangles[id],
+                expectedDeviation = (rotation !== 0 && expected['errors-with-rotation'][id]) || expected.errors[id] || 0;
 
-        const numTriangles = indices.length / 3;
-        assert.ok(numTriangles === expectedTriangles, `${numTriangles} triangles when expected ${expectedTriangles}`);
+            const numTriangles = indices.length / 3;
+            if (rotation === 0) {
+                assert.ok(numTriangles === expectedTriangles, `${numTriangles} triangles when expected ${expectedTriangles}`);
+            }
 
-        if (expectedTriangles > 0) {
-            assert.ok(err <= expectedDeviation, `deviation ${err} <= ${expectedDeviation}`);
-        }
-    });
+            if (expectedTriangles > 0) {
+                assert.ok(err <= expectedDeviation, `deviation ${err} <= ${expectedDeviation}`);
+            }
+        });
+    }
 }
 
 test('infinite-loop', () => {