… test/expected.json
… test/fixtures/touching-holes2.json
… test/fixtures/touching-holes3.json
… test/fixtures/touching-holes4.json
… test/fixtures/touching-holes5.json
… test/fixtures/touching-holes6.json