I've been noticing incorrect triangulations from earcut on real world data recently, even when the input polygons are simple and valid. It seems to happen when polygons contain inner rings that touch at a single vertex.  Here is a screenshot of the issue in the wild:

<img width="931" alt="image" src="https://github.com/user-attachments/assets/8b36b36d-b650-4619-88c1-e23efe25a3a3" />

I reduced the left one to this minimal reproduction case:

![bad](https://github.com/user-attachments/assets/b7f04f32-6c6e-4f74-90c1-038ecfd6f053)

```json
[
  [[0,0],[20,0],[20,25],[0,25],[0,0]],
  [[3,3],[2,12],[9,15],[3,3]],
  [[9,21],[2,12],[7,22],[9,21]]
]
```

Since `eliminateHoles` sorts all holes by their leftmost point, when multiple holes meet at that point if they are processed in the wrong order it makes the border of the new outer polygon shape criss-cross so triangulation can't do the right thing.

This change makes `eliminateHoles` process holes that meet at the same point clockwise so that each ring can be inserted after the end of the one above it.

It fixes these cases:

<table>
<tr>
<td>
before
</td>
<td>
after
</td>
</tr>
<tr>
<td>

![bad](https://github.com/user-attachments/assets/b7f04f32-6c6e-4f74-90c1-038ecfd6f053)

</td>
<td>

![download](https://github.com/user-attachments/assets/31bf6a85-9a58-487f-8e19-94e25f84de83)

</td>
</tr>
<tr>
<td>

![touching-holes3-before](https://github.com/user-attachments/assets/56a3445f-bf42-4da8-9b49-2971fee6ccb3)


</td>
<td>

![touching-holes3-after](https://github.com/user-attachments/assets/85c9d458-311a-4b0c-94c7-6b528897009e)


</td>
</tr>
</table>