@msbarry 2024-12-13
This fixes cases where holes meet at a left-most point, but there still appear to be issues for valid polygons that meet at a right-most point:

But these cases where inner rings meet at a right-most point still seem to fail:

<table>
<tr>
<td>
before
</td>
<td>
after
</td>
</tr>
<tr>
<td>

touching-holes4.json

![image](https://github.com/user-attachments/assets/e32c6910-1100-4485-81a3-cdeaff58ecb0)



</td>
<td>

touching-holes5.json

![image](https://github.com/user-attachments/assets/5d3ef93d-7788-4fa9-96ab-dc5fc4ef3195)

</td>
</tr>
</table>

It appears in these cases that `eliminateHoles` produces a correct outer shell, but then triangulation doesn't treat any of those inner shells as ears because `pointInTriangle` finds one of the vertices of a different triangle at that shared vertex and returns `true`.  I tried changing `pointInTriangle` to return false in those cases where the other point is one of the vertices of the ear. That makes these test cases pass, but causes others to fail:

<table>
<tr>
<td>

![image](https://github.com/user-attachments/assets/6b47e773-c512-4e54-9cc1-08fd47823e9b)

</td>
<td>

![image](https://github.com/user-attachments/assets/8e50ce85-47f3-4307-a518-5b5623851ed0)


</td>

<td>

![image](https://github.com/user-attachments/assets/6df7e853-2a5e-49a6-a235-42fd100cfbef)


</td>
</tr>
</table>

Any ideas what might be needed to fix those cases?