@msbarry 2024-12-14
OK I think I figured out those remaining issues and the test pass now. Let me know what you think.

<table>
<tr>
<td>
touching-holes4.json

![image](https://github.com/user-attachments/assets/be96f14b-33cb-4255-94fc-af6742fac79d)

<td>
<td>
touching-holes5.json

![image](https://github.com/user-attachments/assets/fb85a9af-9c2b-4ca0-a14f-776f96ac7853)

<td>
</tr>
</table>