@msbarry 2024-12-16
Thanks! By "that regression for the simple use case" do you mean the "typical OSM building (15 vertices)" benchmark? Or is it behavior on one of the test cases?