@mourner 2024-12-16
> do you mean the "typical OSM building (15 vertices)" benchmark?

Yep, that one, assuming it's not a fluke result.