@msbarry 2024-12-17
OK got it. I pushed the check inside pointInTriangle but it broke the tests because of the other usages of that function, so I split it into a second function that excludes the first point.  Now they are closer...

this branch:

```
typical OSM building (15 vertices): x 2,785,989 ops/sec ±0.27% (97 runs sampled)
dude shape (94 vertices): x 76,892 ops/sec ±0.26% (101 runs sampled)
dude shape with holes (104 vertices): x 62,760 ops/sec ±0.23% (101 runs sampled)
complex OSM water (2523 vertices): x 1,040 ops/sec ±0.15% (98 runs sampled)
```

and main right afterwards:

```
typical OSM building (15 vertices): x 2,803,361 ops/sec ±0.29% (97 runs sampled)
dude shape (94 vertices): x 77,019 ops/sec ±0.26% (96 runs sampled)
dude shape with holes (104 vertices): x 62,016 ops/sec ±0.99% (97 runs sampled)
complex OSM water (2523 vertices): x 1,024 ops/sec ±0.43% (97 runs sampled)
```