@msbarry 2024-12-18
Ah good call, performance is the same calling pointInTriangle inside pointInTriangleExceptFirst, but less repetetive.

Also this isn't related to my change at all but I noticed there might be some improvement to be had by changing the bbox calculations to use `Math.min(ax, bx, cx)`.

this branch:

```
typical OSM building (15 vertices): x 2,780,054 ops/sec ±0.24% (97 runs sampled)
dude shape (94 vertices): x 78,125 ops/sec ±0.46% (101 runs sampled)
dude shape with holes (104 vertices): x 62,822 ops/sec ±1.06% (100 runs sampled)
complex OSM water (2523 vertices): x 1,081 ops/sec ±0.16% (99 runs sampled)
```

main:

```
typical OSM building (15 vertices): x 2,784,181 ops/sec ±0.44% (94 runs sampled)
dude shape (94 vertices): x 76,417 ops/sec ±1.08% (101 runs sampled)
dude shape with holes (104 vertices): x 62,351 ops/sec ±0.27% (99 runs sampled)
complex OSM water (2523 vertices): x 1,020 ops/sec ±0.34% (96 runs sampled)
```