@mourner 2024-12-18
@msbarry excellent! There were times when `Math.min` / `Math.max` were slower, but v8 is evolving so it's no surprise it's faster now.