@msbarry 2024-12-18
Great! Thanks for the help on this! For reference those benchmarks were all node 21 on a 2021 M1 Macbook Pro.