@mourner 2024-12-25
@msbarry Merry Christmas! Looks like this change produces some weird triangulation failures in GL JS tests (seeing results from https://github.com/mapbox/mapbox-gl-js/pull/13371, e.g. here https://output.circle-artifacts.com/output/job/d76e512a-1057-4c34-9a84-c32415369555/artifacts/0/test/integration/render-tests/index.html) — will need investigation. If it's not an easy thing to fix, we'll probably have to hold off on upgrading and potentially revert...