@msbarry 2024-12-25
Are you able to reproduce those failures as test cases in earcut? If so I can take a look in a few days.