@msbarry 2025-01-21
Hey @mourner just checking in on this, any update reproducing those failed test cases in earcut?